requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-telegram-bot==20.6
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
alembic==1.12.1
asyncpg==0.28.0
apscheduler==3.10.4
pillow==10.1.0
python-multipart==0.0.6 